The first REAL commit.
Adds the basic code for getting slick to work, as well as some basic
classes I wrote such as Vector2d and IsoTile (which is nowhere close to
being done).

Added images of each tank in all 4 diagonal orientations, plus a basic
unit for each side (which is the same on all 4 sides, no need to render
that out). Also added image for the basic ground, I'm really feeling a
'Tron' vibe, so DarkBaseSmall.png is what the tiles will be based on.